+++ lib/braintree/transaction.rb
@@ -149,6 +149,7 @@ module Braintree
     attr_reader :retrieval_reference_number
     attr_reader :risk_data
     attr_reader :samsung_pay_card_details
+    attr_reader :sca_exemption_requested
     attr_reader :service_fee_amount
     attr_reader :settlement_batch_id
     attr_reader :shipping_amount
@@ -308,6 +309,7 @@ module Braintree
       @us_bank_account_details = UsBankAccountDetails.new(attributes[:us_bank_account]) if attributes[:us_bank_account]
       @visa_checkout_card_details = VisaCheckoutCardDetails.new(attributes[:visa_checkout_card])
       @samsung_pay_card_details = SamsungPayCardDetails.new(attributes[:samsung_pay_card])
+      @sca_exemption_requested = attributes[:sca_exemption_requested]
       authorization_adjustments.map! { |attrs| AuthorizationAdjustment._new(attrs) } if authorization_adjustments
 
       installments.map! { |attrs| Installment.new(attrs) } if installments